int to long

+++ app/src/main/java/com/flatstack/android/models/User.java
@@ -30,7 +30,7 @@ public class User {
 		return _id;
 	}
 
-	public void set_id(int _id) {
+	public void set_id(long _id) {
 		this._id = _id;
 	}
